+++ app/(service)/(my)/my-study/completed/page.tsx
@@ -2,12 +2,12 @@
 
 import { Plus } from 'lucide-react';
 import React, { useState } from 'react';
+import Pagination from '@/components/ui/pagination';
 import { MemberStudyItem } from '@/features/study/group/api/group-study-types';
 import { useMemberStudyListQuery } from '@/features/study/group/model/use-member-study-list-query';
 import CompletedGroupStudyList from '@/features/study/group/ui/completed-group-study-list';
 import OpenGroupStudyModal from '@/features/study/group/ui/open-group-modal';
 import Button from '@/shared/ui/button';
-import Pagination from '@/shared/ui/pagination';
 
 interface MemberGroupStudyList extends MemberStudyItem {
   type: 'GROUP_STUDY';
