+++ src/features/admin/ui/member-list-table.tsx
@@ -2,13 +2,13 @@
 
 import { useRouter } from 'next/navigation';
 import { useEffect, useRef, useState } from 'react';
+import Pagination from '@/components/ui/pagination';
 import { useDebounce } from '@/shared/lib/debounce';
 import { formatYYYYMMDD } from '@/shared/lib/time';
 import Badge from '@/shared/ui/badge';
 import Button from '@/shared/ui/button';
 import Checkbox from '@/shared/ui/checkbox';
 import { SingleDropdown } from '@/shared/ui/dropdown';
-import Pagination from '@/shared/ui/pagination';
 import FilledX from 'public/icons/filled-x.svg';
 import SealCheckIcon from 'public/icons/seal-check.svg';
 import SearchIcon from 'public/icons/search.svg';
